pyproject.toml
# pyproject.toml
[tool.poetry]
name = "star-competency-app"
version = "0.1.0"
description = "Application for STAR method stories and case study analysis aligned with company competencies"
authors = ["Muhammad Samy <azzam.ma@nahdi.sa>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.5"
anthropic = "^0.9.0"
msal = "^1.24.0"
pillow = "^9.5.0"
pydantic = "^2.4.0"
python-dotenv = "^1.0.0"
flask-login = "^0.6.2"
gunicorn = "^21.2.0"
flask-sqlalchemy = "^3.1.0"
alembic = "^1.11.0"
bleach = "^5.0.1"
marshmallow = "^3.19.0"
openai = "^1.74.0"
pydantic-settings = "^2.8.1"
pytesseract = "^0.3.13"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "star_competency_app.main:main"
